refactor: modularize socket event handling into separate action controllers
