+++ server/tests/ci-smoke.test.js
@@ -1,5 +1,5 @@
 describe("CI Smoke Test", () => {
-  it("should pass basic sanity checks", () => {
+  it("should pass basic sanity checks on the server", () => {
     expect(true).toBe(true);
   });
 });
